Move default parameters outside of app package
