+++ src/aiidalab_qe/app/configuration/advanced/general/model.py
@@ -4,8 +4,8 @@ import typing as t
 
 import traitlets as tl
 
-from aiidalab_qe.app.parameters import DEFAULT_PARAMETERS
 from aiidalab_qe.common.panel import PanelModel
+from aiidalab_qe.parameters import DEFAULT_PARAMETERS
 
 DEFAULT = t.cast(dict, DEFAULT_PARAMETERS)
