+++ src/aiidalab_qe/app/configuration/model.py
@@ -3,13 +3,13 @@ from __future__ import annotations
 import traitlets as tl
 
 from aiida_quantumespresso.common.types import RelaxType
-from aiidalab_qe.app.parameters import DEFAULT_PARAMETERS
 from aiidalab_qe.common.mixins import (
     HasInputStructure,
     HasModels,
 )
 from aiidalab_qe.common.panel import PanelModel
 from aiidalab_qe.common.wizard import QeConfirmableWizardStepModel
+from aiidalab_qe.parameters import DEFAULT_PARAMETERS
 
 DEFAULT: dict = DEFAULT_PARAMETERS  # type: ignore
