+++ src/aiidalab_qe/app/result/components/summary/model.py
@@ -10,10 +10,10 @@ from jinja2 import Environment
 from aiida import orm
 from aiida.cmdline.utils.common import get_workchain_report
 from aiida_quantumespresso.workflows.pw.bands import PwBandsWorkChain
-from aiidalab_qe.app.parameters import DEFAULT_PARAMETERS
 from aiidalab_qe.app.result.components import ResultsComponentModel
 from aiidalab_qe.app.static import styles, templates
 from aiidalab_qe.common.time import format_time, relative_time
+from aiidalab_qe.parameters import DEFAULT_PARAMETERS
 from aiidalab_qe.utils import get_pseudo_info
 
 DEFAULT: dict = DEFAULT_PARAMETERS  # type: ignore
