+++ src/aiidalab_qe/app/submission/__init__.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 
 import ipywidgets as ipw
 
-from aiidalab_qe.app.parameters import DEFAULT_PARAMETERS
 from aiidalab_qe.app.utils import get_entry_items
 from aiidalab_qe.common.code import PluginCodes, PwCodeModel
 from aiidalab_qe.common.infobox import InAppGuide
@@ -19,6 +18,7 @@ from aiidalab_qe.common.panel import (
 from aiidalab_qe.common.setup_codes import QESetupWidget
 from aiidalab_qe.common.widgets import LinkButton
 from aiidalab_qe.common.wizard import QeConfirmableDependentWizardStep
+from aiidalab_qe.parameters import DEFAULT_PARAMETERS
 
 from .global_settings import GlobalResourceSettingsModel, GlobalResourceSettingsPanel
 from .model import SubmissionStepModel
